More future-proofing of indices

+++ PSF_Testing/psf_testing/stacking.py
@@ -30,7 +30,7 @@ from psf_testing.moments.centre_image import centre_image
 from psf_testing.moments.estimate_background import get_background_level
 
 
-def make_stacks(stars, stack_size=(2 * mv.default_weight_rmax + 1),
+def make_stacks(stars, stack_size=int(2 * mv.default_weight_rmax + 1),
                 weight_func=mv.default_prim_weight_func):
 
     stacks = {}
